Attempt to get chemkinimporter running QM and PDep
Not sure this is a good idea, but my hope is that we will include
some well-skipping reactions in our model, which will match with
things other people have in their models. i.e. many of the "unmatched"
reactions in models we are importing are not elementary.

Our current pruning of species that can't match something in the model might
prevent us from discovering some pathways. eg.
A + B -> C -> D + E
but their model does not include anything as big as C, we will delete C and
so may never discover the reaction
A + B -> D + E

The QM should help us get better thermo.