Can now quit when exhausted instead of waiting for input from web front end.
Either call with the --quit_when_exhausted command line option, or set the
RMG_QUIT_WHEN_EXHAUSTED environment variable before you begin.

This could be useful if you want to batch-start a bunch of jobs just to regenerate
the thermo.py and kinetics.py files, without all the jobs waiting for ever for input.
They will run as far as they can, then quit.
You could, for example, try this:

    $ export RMG_QUIT_WHEN_EXHAUSTED=True
    $ find . -name import.sh -execdir sh -c "./import.sh &" \;

or, on a cluster with LSF:

    $ export RMG_QUIT_WHEN_EXHAUSTED=True
    $ find . -name import.sh -execdir sh -c "bsub -q ht-10g  < import.sh" \;