Stop segfault due to None atomtype on Argon
This causes a horrid crash because the VF2 algorithm tries to
see if Ar is subgraph isomorphic with a reaction template, but
that typecasts AtomType to the atom type, but for Argon the atom
type is None, which isn't an AtomType, so it crashes.

We now make Argon inert, and remove it from the reaction list.
Bit of a pain.