Don't complain about unmarked duplicate reactions.
RMG thinks that pdep and non-pdep versions of a reaction
should be marked as duplicates, i.e. A(+M) is equivalent to A
but CHEMKIN doesn't. 
For now, just mark everything that RMG wants to be a duplicate,
as a duplicate, even if it wasn't marked as such by chemkin.