Caching the species' chemical formula to speed up the importer.
This has a hazard that if the molecule is updated, the species's chemical
formula is cached and will not be updated. However, this will not break
any current code, because this attribute is new and thus no code is using
it. Also, it is already a hazard that exists for fingerprint, inchi, and smiles
which are treated this way.

Will require a rebuild of cython code.